scafolding e comentarios ao codigo e mensagem de erro personalizadas

+++ Books/Controllers/AuthorsController.cs
@@ -14,19 +14,19 @@ namespace Books.Controllers
     {
         private readonly BooksDbContext _context;
 
-        public AuthorsController(BooksDbContext context)
+        public AuthorsController(BooksDbContext context)//recebe a bd dos books
         {
             _context = context;
         }
 
         // GET: Authors
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index()//lista dos autores
         {
-            return View(await _context.Author.ToListAsync());
+            return View(await _context.Author.ToListAsync());//alteracoes assincronas
         }
 
         // GET: Authors/Details/5
-        public async Task<IActionResult> Details(int? id)
+        public async Task<IActionResult> Details(int? id)//recebe o id do autor que é opcional
         {
             if (id == null)
             {
@@ -34,9 +34,10 @@ namespace Books.Controllers
             }
 
             var author = await _context.Author
-                .FirstOrDefaultAsync(m => m.AuthorId == id);
+                .SingleOrDefaultAsync(m => m.AuthorId == id);// um registo ou o default que é null
             if (author == null)
             {
+                //todo: talvez alguem tenha apagado esse autor. mostrar uma mensagem apropriada para o user
                 return NotFound();
             }
 
@@ -53,13 +54,15 @@ namespace Books.Controllers
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("AuthorId,Name")] Author author)
+        [ValidateAntiForgeryToken]//validacao de seguranca
+        public async Task<IActionResult> Create([Bind("AuthorId,Name")] Author author)//serve para evitar alguns ataques, só recebo os campos que estão no bind 
         {
             if (ModelState.IsValid)
             {
+                //todo: validacoes adicionais antes de inserir o autor
                 _context.Add(author);
                 await _context.SaveChangesAsync();
+                //todo: informar o user, autor criado com sucesso
                 return RedirectToAction(nameof(Index));
             }
             return View(author);
@@ -76,6 +79,7 @@ namespace Books.Controllers
             var author = await _context.Author.FindAsync(id);
             if (author == null)
             {
+                //todo: talvez alguem tenha apagado esse autor. mostrar uma mensagem apropriada para o user
                 return NotFound();
             }
             return View(author);
@@ -104,10 +108,13 @@ namespace Books.Controllers
                 {
                     if (!AuthorExists(author.AuthorId))
                     {
+                        // todo: talvez alguem apagou isto
+                        //informar o user se quer criar um novo com os mesmos dados
                         return NotFound();
                     }
                     else
                     {
+                        // todo: mostrar o erro e perguntar se quer tentar outra vez
                         throw;
                     }
                 }
@@ -128,6 +135,7 @@ namespace Books.Controllers
                 .FirstOrDefaultAsync(m => m.AuthorId == id);
             if (author == null)
             {
+                // todo: talvez alguem apagou isto, informar o user
                 return NotFound();
             }
 
@@ -140,8 +148,11 @@ namespace Books.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var author = await _context.Author.FindAsync(id);
+            if(author != null) { 
             _context.Author.Remove(author);
             await _context.SaveChangesAsync();
+            }
+            // todo: informar o user que o autor foi apagado com sucesso
             return RedirectToAction(nameof(Index));
         }
